+++ CHANGELOG.md
@@ -10,6 +10,17 @@
 ## 1.0.14-preview (2017-08-31)
 
 ### Bug Fixes
+- String template issue fixed.
+- Exception at clearTemplate during Component destroy is resolved.
+
+## 1.0.16-preview (2017-09-08)
+
+### Bug Fixes
+- Model value not working with checkbox issue fixed.
+
+## 1.0.14-preview (2017-08-31)
+
+### Bug Fixes
 - Rollup bundling issue fixed.
 
 ## 1.0.11-preview (2017-08-03)